+++ docs/changelog.md
@@ -7,6 +7,7 @@
 - Add smoke test for workflowish overall
 - Fixed unindent bug for symlinks
 - Make sure symlinked node is passed to update cycle
+- Smaller symlink
 
 # Version 3.3.0: Miscellaneous Bugfixes 
 - Add workflowish item creation in script engine